myproject/requirement.txt
Requirement:
django-webpack-loader-0.5.0
